Added test for show hide custom date range picker
Split the tests into two files to keep the mocks isolated. Scoping mocks within one file using jest.doMock() is possible, but is complicated for ES6 imports. Separating the files unfortunately seems to be the most straightforward solution.
